refactor : 변수명, 메서드명 리팩토링
클래스명, 매개변수 등과 이어졌을 때 중복이 없도록 리팩토링
