+++ src/test/java/lotto/model/lotto/PurchaseAmountTest.java
@@ -3,7 +3,7 @@ package lotto.model.lotto;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-import lotto.exception.Exceptions;
+import lotto.common.Exceptions;
 import lotto.model.purchaseAmount.PurchaseAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
