+++ src/test/java/lotto/model/winningNumber/BonusNumberGeneratorTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import lotto.exception.Exceptions;
+import lotto.common.Exceptions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
